Added MOV_MtoRD
Turns out the writing register data to RAM was working correctly, but we had the write enable signal as ram_wire instead of ram_write.

Also, added the function to read from RAM to send data to a register. Managed to optimize the operation to only use 3 clock cycles instead of 5.